move `TransactionsNRInfo` into `HistoryHeader`
