+++ components/history/Header.vue
@@ -23,18 +23,50 @@
       />
     </div>
     <ExamineNamesList class="pl-4" :choices="nr.names" highlight />
-    <TransactionsNRInfo :data="nr" class="w-full" />
+
+    <div class="flex w-full text-sm">
+      <div class="grid grow grid-cols-4">
+        <span class="font-bold">Submitted Date</span>
+        <span>{{ submittedDate }}</span>
+
+        <span class="font-bold">Expiry Date</span>
+        <span>
+          {{ expirationDate }}
+        </span>
+
+        <span class="font-bold">Request Status</span>
+        <span>{{ statusDisplay }}</span>
+
+        <span class="font-bold">Consent</span>
+        <span>{{ consentDisplay }}</span>
+
+        <span class="font-bold">Additional Information</span>
+        <span>{{ nr.additionalInfo }}</span>
+
+        <span v-if="includeRecipient" class="font-bold">Recipient</span>
+        <span v-if="includeRecipient">{{ nr.applicants.emailAddress }}</span>
+      </div>
+    </div>
   </div>
 </template>
 
 <script setup lang="ts">
 import type { NameRequest } from '~/types'
 import NameRequestTypeInfo from '../NameRequestTypeInfo.vue'
-import { getRequestTypeDisplay } from '~/util/request-type'
+import { ConsentFlag } from '~/enums/codes'
+import { Status } from '~/enums/nr-status'
+import { parseDate, getFormattedDateWithTimeAndZone } from '~/util/date'
+import {
+  getRequestTypeDisplay,
+  getStatusDisplay,
+  getConsentDisplay,
+} from '~/util/display-format'
 
 const { nr } = defineProps<{
   nr: NameRequest
   loading?: boolean
+  /** Include the `recipient` field in the NR info of the header. */
+  includeRecipient?: boolean
 }>()
 
 const isPriority = computed(() => nr.priorityCd === 'Y')
@@ -45,5 +77,23 @@ const requestTypeDisplay = computed(() =>
     nr.request_action_cd,
     nr.entity_type_cd
   )
+)
+
+const submittedDate = computed(() =>
+  getFormattedDateWithTimeAndZone(parseDate(nr.submittedDate))
+)
+
+const expirationDate = computed(() =>
+  nr.expirationDate
+    ? getFormattedDateWithTimeAndZone(parseDate(nr.expirationDate))
+    : 'N/A'
+)
+
+const statusDisplay = computed(() =>
+  getStatusDisplay(nr.stateCd as Status, nr.names)
+)
+
+const consentDisplay = computed(() =>
+  getConsentDisplay(nr.consent_dt, nr.consentFlag || ConsentFlag.Waived)
 )
 </script>
