+++ components/transactions/Entry.vue
@@ -44,7 +44,11 @@ import { DateTime } from 'luxon'
 import { useTransactions } from '~/store/transactions'
 import type { TransactionEntry } from '~/types'
 import { getFormattedDateWithTimeAndZone, parseDate } from '~/util/date'
-import { getRequestTypeDisplay } from '~/util/request-type'
+import {
+  getRequestTypeDisplay,
+  getStatusDisplay,
+  getConsentDisplay,
+} from '~/util/display-format'
 
 const { entry } = defineProps<{
   entry: TransactionEntry
@@ -57,7 +61,7 @@ const eventDate = computed(() =>
 )
 
 const statusDisplay = computed(() =>
-  transactions.getStatusDisplay(entry.stateCd, entry.names)
+  getStatusDisplay(entry.stateCd, entry.names)
 )
 
 const expiryDate = computed(() => {
@@ -86,7 +90,7 @@ const requestTypeDisplay = computed(() =>
 )
 
 const consentDisplay = computed(() =>
-  transactions.getConsentDisplay(entry.consent_dt, entry.consentFlag)
+  getConsentDisplay(entry.consent_dt, entry.consentFlag)
 )
 
 const isStaffCommentAction = computed(
