+++ store/transactions.ts
@@ -1,7 +1,4 @@
-import { ConsentFlag } from '~/enums/codes'
-import { Status } from '~/enums/nr-status'
 import {
-  type NameChoice,
   type NameRequest,
   type TransactionEntry,
   type Transactions,
@@ -47,40 +44,11 @@ export const useTransactions = defineStore('transactions', () => {
     }
   }
 
-  function getStatusDisplay(status: Status | null, names: Array<NameChoice>) {
-    if (!status) return 'N/A'
-
-    let displayState = status as string
-    if (status === Status.Conditional) {
-      displayState = 'CONDITIONAL APPROVED'
-    } else if (status === Status.Consumed && names && names.length > 0) {
-      const approvedName = names.find((name) =>
-        [Status.Approved, Status.Condition].includes(name.state)
-      )
-      displayState =
-        approvedName?.state === Status.Condition
-          ? 'CONDITIONAL APPROVED'
-          : 'APPROVED'
-      displayState += ` / Used for ${approvedName?.corpNum}`
-    }
-    return displayState
-  }
-
-  function getConsentDisplay(date: string | null, consentFlag: ConsentFlag) {
-    return date
-      ? 'Required. Received.'
-      : consentFlag === ConsentFlag.Required
-      ? 'Required. Not yet received.'
-      : 'Not required'
-  }
-
   return {
     nr,
     transactions,
     loadingNr,
     loadingTransactions,
     initialize,
-    getStatusDisplay,
-    getConsentDisplay,
   }
 })
